services/user-service/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
prometheus-client==0.21.0
python-jose==3.3.0
